+++ example/release_updater_example.dart
@@ -4,12 +4,21 @@ import 'package:release_updater/release_updater_io.dart';
 
 void main() async {
   var storage = ReleaseStorageDirectory('appx', Directory('/install/path'));
+
   var provider =
       ReleaseProviderHttp.baseURL('https://your.domain/appx/releases');
 
   var releaseUpdater = ReleaseUpdater(storage, provider);
 
-  var version = await releaseUpdater.update();
+  print('-- Updating...');
+  var updatedToVersion = await releaseUpdater.update();
 
-  print('-- Updated to version: $version');
+  if (updatedToVersion != null) {
+    print('-- Updated to version: $updatedToVersion');
+  }
+
+  var runExecutable = await storage.currentReleaseFilePath('run.exe');
+
+  print('-- Running: $runExecutable');
+  Process.run(runExecutable!, []);
 }