+++ lib/release_updater_io.dart
@@ -2,4 +2,5 @@
 library release_updater.io;
 
 export 'release_updater.dart';
+export 'src/release_updater_io.dart';
 export 'src/release_updater_storage_io.dart';