+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 1.0.2
 
 - Added `currentReleaseFilePath` and `currentReleaseFile`.
+- When loading a `ReleaseBundle` from a Zip, detected executables by file extension.
+- When storing a `ReleaseFile`, ensure that the `executable` permission is set.
+- Added `startReleaseProcess` and `runReleaseProcess`.
 
 ## 1.0.1